Add `get_or_insert` method.
Return the value assigned to the given key, if present.

If the key is not present, the given yield block will be executed to
provide a value to assign to the key, which will also be returned.

If the key was already present, the yield block will not be executed.