Remove `*_until`-style iteration methods.
This style/convention is a legacy from an old version of the Savi
language, in which yield blocks could not be "jumped out of".

Now that `return`, `error!`, and `break` can all jump out of
a yield block, the pattern (of allowing the caller to yield back
a `Bool` to indicate whether iteration should continue) is obsolete,
and should be removed to reduce complexity and maintenance burden.